Add a confirmation dialog while resetting settings
