+++ lib/features/table_of_contents/chapter_list/toc_sliver_chapter_list_item.dart
@@ -2,13 +2,13 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 
-import '../../../data/bookmark_data.dart';
 import '../../../data/chapter_data.dart';
 import '../../../processor/bookmark_processor.dart';
 import '../../common_components/common_loading.dart';
-import '../../reader/reader.dart';
+import '../../common_components/draggable_feedback_widget.dart';
+import '../../common_components/draggable_placeholder_widget.dart';
 import '../bloc/toc_bloc.dart';
-import 'toc_chapter_title.dart';
+import '../widgets/toc_chapter_widget.dart';
 
 class TocSliverChapterListItem extends StatelessWidget {
   final ChapterData chapterData;
@@ -21,7 +21,6 @@ class TocSliverChapterListItem extends StatelessWidget {
     final double verticalPadding = MediaQuery.of(context).textScaler.scale(fontSize) / 2;
     final int chapterNumber = chapterData.ordinalNumber;
     final String bookName = chapterData.bookName;
-    final double iconSize = IconTheme.of(context).size ?? 24.0;
 
     return Padding(
       padding: EdgeInsets.symmetric(horizontal: 16.0, vertical: verticalPadding),
@@ -66,88 +65,23 @@ class TocSliverChapterListItem extends StatelessWidget {
                 });
               },
               data: chapterData,
-              feedback: Container(
+              feedback: DraggableFeedbackWidget(
                 width: constraints.maxWidth,
-                decoration: BoxDecoration(
-                  color: Theme.of(context).colorScheme.surface.withOpacity(0.9),
-                  borderRadius: BorderRadius.circular(16.0),
-                  boxShadow: [
-                    BoxShadow(
-                      color: Theme.of(context).colorScheme.shadow.withOpacity(0.2),
-                      blurRadius: 8.0,
-                      spreadRadius: 0.0,
-                      offset: const Offset(0.0, 4.0),
-                    ),
-                  ],
-                ),
-                child: _chapterWidget(context, chapterData, isBookmarked),
+                child: TocChapterWidget(chapterData: chapterData, isBookmarked: isBookmarked),
               ),
-              childWhenDragging: Opacity(
-                opacity: 0.3,
-                child: Container(
-                  width: constraints.maxWidth,
-                  color: Theme.of(context).colorScheme.surface,
-                  child: _chapterWidget(context, chapterData, isBookmarked),
-                ),
+              childWhenDragging: DraggablePlaceholderWidget(
+                width: constraints.maxWidth,
+                child: TocChapterWidget(chapterData: chapterData, isBookmarked: isBookmarked),
               ),
               child: Container(
                 width: constraints.maxWidth,
                 color: Theme.of(context).colorScheme.surface,
-                child: _chapterWidget(context, chapterData, isBookmarked),
+                child: TocChapterWidget(chapterData: chapterData, isBookmarked: isBookmarked),
               ),
             );
           },
         );
       }),
-    );
-  }
-
-  Route _navigateToReader(String bookName, int chapterNumber) {
-    return PageRouteBuilder(
-      pageBuilder: (context, animation, secondaryAnimation) => ReaderWidget(bookName, chapterNumber),
-      transitionDuration: const Duration(milliseconds: 300),
-      reverseTransitionDuration: const Duration(milliseconds: 300),
-      transitionsBuilder: (context, animation, secondaryAnimation, child) {
-        return FadeTransition(
-          opacity: Tween<double>(begin: 0.0, end: 1.0).animate(
-            CurvedAnimation(
-              parent: animation,
-              curve: Curves.easeInOut,
-            ),
-          ),
-          child: ScaleTransition(
-            scale: Tween<double>(begin: 0.0, end: 1.0).animate(
-              CurvedAnimation(
-                parent: animation,
-                curve: Curves.easeInOut,
-              ),
-            ),
-            child: child,
-          ),
-        );
-      },
-    );
-  }
-
-  Widget _chapterWidget(BuildContext context, ChapterData chapterData, bool isBookmarked) {
-    final int chapterNumber = chapterData.ordinalNumber;
-    final String bookName = chapterData.bookName;
-    final double iconSize = IconTheme.of(context).size ?? 24.0;
-    return TextButton.icon(
-      onPressed: () => Navigator.of(context)
-          .push(_navigateToReader(bookName, chapterNumber))
-          .then((_) => BlocProvider.of<TocCubit>(context).refresh()),
-      style: TextButton.styleFrom(
-        alignment: Alignment.centerLeft,
-        shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(16.0)),
-        foregroundColor: Theme.of(context).colorScheme.onSurface,
-      ),
-      icon: Icon(
-        isBookmarked ? Icons.bookmark_rounded : Icons.numbers_rounded,
-        size: MediaQuery.of(context).textScaler.scale(iconSize),
-        color: isBookmarked ? Theme.of(context).colorScheme.error : null,
-      ),
-      label: TocChapterTitle(chapterData),
     );
   }
 }
